+++ lib/common/StreamableMemBlock.h
@@ -30,7 +30,7 @@ public:
 	~StreamableMemBlock();
 	
 	void Set(const StreamableMemBlock &rBlock);
-	void Set(void *pBuffer, int Size);
+	void Set(void *pBuffer, size_t Size);
 	void Set(IOStream &rStream, int Timeout);
 	StreamableMemBlock &operator=(const StreamableMemBlock &rBlock)
 	{
@@ -46,7 +46,7 @@ public:
 	void *GetBuffer() const;
 	
 	// Size of block
-	int GetSize() const {return mSize;}
+	size_t GetSize() const {return mSize;}
 
 	// Buffer empty?
 	bool IsEmpty() const {return mSize == 0;}
@@ -59,12 +59,12 @@ public:
 	void ResizeBlock(int Size);
 
 protected:	// be careful with these!
-	void AllocateBlock(int Size);
+	void AllocateBlock(size_t Size);
 	void FreeBlock();
 
 private:
 	void *mpBuffer;
-	int mSize;
+	size_t mSize;
 };
 
 #endif // STREAMABLEMEMBLOCK__H
