+++ lib/compress/CompressStream.h
@@ -32,8 +32,8 @@ private:
 	CompressStream &operator=(const CompressStream &);
 public:
 
-	virtual int Read(void *pBuffer, int NBytes, int Timeout = IOStream::TimeOutInfinite);
-	virtual void Write(const void *pBuffer, int NBytes);
+	virtual size_t Read(void *pBuffer, size_t NBytes, int Timeout = IOStream::TimeOutInfinite);
+	virtual void Write(const void *pBuffer, size_t NBytes);
 	virtual void WriteAllBuffered();
 	virtual void Close();
 	virtual bool StreamDataLeft();
